Support intra-page URL fragment identifiers
Allows you to link directly to, e.g., an individual attribute or method
in a specific doc page.
Supports `<core-doc-viewer>`s with multiple `sources=` set, or with a
single `url=` set.
Maintains compatibility with using class names as the URL fragment
identifier to switch between classes when multiple `sources=` is set.