Scroll to top if data changes
The problem: 
When reusing core-doc-page for multiple docs, scroll position is preserved when docs change. It is a bad user experience, you end up looking at the middle of a freshly loaded doc.
The fix:
Scroll to top when user data changes.
Question:
Is there a way to do this without traversing core-header-panel's internal structure? Hate doing that...